pyproject.toml
[project]
name = "auto-content-pipeline"
version = "0.1.0"
description = "Automated data collection, processing, classification, and GitHub sync pipeline"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "feedparser>=6.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "selectolax>=0.3",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
pipeline = "src.main:main"
